feat(bar_stock): T1.9 — INTEGER -> NUMERIC(12, 3) consistency pass
Closes the last Tier 1 item from docs/roadmap.md. Brings bar_stock
quantity columns in line with the rest of the schema. Tier 1: 9/9 done.

WHY:
  Every other quantity column in the system was already NUMERIC:
    stock_transactions.qty           NUMERIC(12, 3)
    warehouse_inventory.current_qty  NUMERIC(12, 2)
    warehouse_allocations.reserved_qty  NUMERIC(12, 2)
    warehouse_scans.qty              NUMERIC(12, 2)

  bar_stock was the outlier. allocated_qty / current_qty / returned_qty
  were all INTEGER. This silently truncated fractional consumption when
  the cascade engine wrote bar_stock from stock_transactions:
    - tx records 0.5 bottles consumed
    - bar_stock.current_qty rounds to nearest int
    - reconciliation reports show ledger-vs-snapshot drift

  Stops happening as of this commit. NUMERIC(12, 3) matches
  stock_transactions.qty so arithmetic between them is now lossless.

MIGRATION (k1_alter_bar_stock_to_numeric):
  ALTER COLUMN ... TYPE NUMERIC(12, 3) USING (col::numeric(12, 3))
  for all 3 columns. Postgres handles INT -> NUMERIC widening cleanly,
  no data loss, no row scan needed.

  Five existing CHECK constraints survive automatically:
    bar_stock_allocated_nonneg, bar_stock_current_nonneg,
    bar_stock_returned_nonneg, bar_stock_current_lte_allocated,
    bar_stock_returned_lte_allocated.

  Downgrade casts back to INTEGER (would silently truncate any fractional
  values that landed during the NUMERIC window — acceptable since
  downgrade is explicit).

ORM (bar_stock/models.py):
  All 3 columns now Mapped[Decimal] with Numeric(12, 3).
  Numeric added to sqlalchemy imports. Decimal added to typing imports.
  returned_qty kept its server_default=text('0') verbatim.

PYDANTIC (stock_transactions/schemas.py):
  3 fields switched int -> Decimal. The model_config json_encoders={
  Decimal: float} was already present (left in place — Pydantic's float
  encoder loses precision but is fine for API response wire format
  where the cascade engine has already locked in the truth).

ALERTS ENGINE (alerts/engine.py):
  Type hints widened: 'current_qty: int | None' -> 'Decimal | int | None'.
  No runtime change — Python is duck-typed for arithmetic, ==, and
  format strings on Decimal vs int. The annotations now describe what
  the code actually accepts.

VERIFIED END-TO-END:
  alembic upgrade head  -> ran clean, single transaction
  psql \d bar_stock      -> all 3 cols numeric(12,3)
  SQLAlchemy introspect -> all 3 NUMERIC(12, 3)
  Live data read        -> 4 rows, values Decimal('75.000') etc.,
                           Python types match DB + ORM.

Tier 1 progress: 9 of 9 complete.

Ref: docs/roadmap.md T1.9.

+++ backend/app/modules/stock_transactions/schemas.py
@@ -150,9 +150,9 @@ class ReconciliationLine(BaseModel):
     bar_id: UUID
     product_id: UUID
     bar_stock_id: UUID
-    allocated_qty: int
-    current_qty: int
-    returned_qty: int
+    allocated_qty: Decimal
+    current_qty: Decimal
+    returned_qty: Decimal
     expected_consumption: Decimal
     actual_consumption: Decimal
     anomaly_qty: Decimal
